requirements.txt
albumentations==2.0.5
fastapi==0.128.0
matplotlib==3.10.8
numpy==2.4.1
opencv_python==4.11.0.86
opencv_python_headless==4.11.0.86
Pillow==12.1.0
scipy==1.17.0
segmentation_models_pytorch==0.4.0
skimage==0.0
torch==2.2.2
torchmetrics==1.7.1
torchvision==0.17.2
tqdm==4.67.1
uvicorn==0.40.0
